[webview_flutter] 完成flutter webview 进度值回调Android端实现

+++ packages/webview_flutter/lib/src/webview_method_channel.dart
@@ -44,7 +44,7 @@ class MethodChannelWebViewPlatform implements WebViewPlatformController {
         _platformCallbacksHandler.onPageStarted(call.arguments['url']);
         return null;
       case 'onProgressChanged':
-        _platformCallbacksHandler.onProgressChanged(call.arguments['progress']);
+        _platformCallbacksHandler.onProgressChanged(call.arguments['progress'] as double);
         return null;
       case 'onWebResourceError':
         _platformCallbacksHandler.onWebResourceError(
